1）新增app 前后台切换事件 上报 app_state_changed 2）新增通用属性 is_foreground 区分应用在前后台

+++ roiquery-core/src/main/java/com/roiquery/analytics/data/db/EventDataOperation.kt
@@ -278,7 +278,7 @@ internal class EventDataOperation(private val mContext: Context) {
             AnalyticsImp.getInstance(context)?.maxCacheSize!!
         } catch (e: Exception) {
             LogUtils.printStackTrace(e)
-            32 * 1024 * 1024
+            (32 * 1024 * 1024).toLong()
         }
     }
